[es6][template strings] change return value for function call
While the test will work fine on tddbin.com, that is a hard dependency
on the domain to run the tests. This change removes that hard
dependency and uses a string return value so that the tests can be run
even from the command line and still get it to pass.