[es6][rest] change use of `arguments` keyword in `eliminate arguments` test
the use of the `arguments` keyword makes mocha to hang indefinitely
without exit in case of the failure condition. This probably could
be because `arguments` is a reserved keyword and is not available in
arrow functions, but node / mocha is trying to access / execute it.
The change gives a more obvious hint to the solution, but would avoid
the indefinite hang of the test in the failure scenario.